+++ test/example.test.js
@@ -1,17 +1,19 @@
-// // IMPORT MODULES under test here:
-// // import example from '../src/example.js';
-// import isYes from './example.test.js';
+import isYes from './example.test.js';
 
-// const test = QUnit.test;
+const test = QUnit.test; 
 
-// test('time to test a function', function(assert) {
-//     //Arrange
-//     // Set up your parameters and expectations
+test('yes, y, yasss is yes', function(assert) {
+    const result = isYes('Y');
+    const result1 = isYes('YES');
+    const result2 = isYes('yes');
+    const result3 = isYes('uhhuh');
+    const result4 = isYes('yAsSs');
 
-//     //Act 
-//     // Call the function you're testing and set the result to a const
+    // Assert 
+    assert.equal(result, true);
+    assert.equal(result1, true);
+    assert.equal(result2, true);
+    assert.equal(result3, false);
+    assert.equal(result4, false);
 
-//     //Assert
-//     // Make assertions about what is expected valid result
-//     assert.equal(true, false);
-// });
+});
