+++ test/index.html
@@ -40,7 +40,7 @@
       value: storageMock(),
     });</script>
 
-  <script type="module" src="tests.js"></script>
+  <script type="module" src="example.test.js"></script>
 </body>
 
 </html>
